% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.selected.R
\name{feature.selected}
\alias{feature.selected}
\title{Features which are used to build a tree}
\usage{
feature.selected(tree)
}
\arguments{
\item{tree}{:Tree structure made by getTree() function}
}
\value{
An array of characters which is the name of those covariate used in the tree
}
\description{
The name of covariates which use to partition the covariate space.
}
\details{
The sequence only contain unique name(Only count one time even the single covariate use to split twice or more)
}
\examples{
data(Lung,package="compound.Cox")
train_Lung=Lung[which(Lung[,"train"]==TRUE),] #select training data
t.vec=train_Lung[,1]
d.vec=train_Lung[,2]
x.mat=train_Lung[,-c(1,2,3)]
res=uni.tree(t.vec,d.vec,x.mat,P.value=0.01,d0=0.01,S.plot=FALSE,score=TRUE)
feature.selected(res)
}
