% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.classification.R
\name{risk.classification}
\alias{risk.classification}
\title{Risk of rank generated by tree-based method}
\usage{
risk.classification(tree, X.mat)
}
\arguments{
\item{tree}{:Tree structure made by uni.tree() function}

\item{X.mat}{:n by p matrix of covariates, where n is the sample size and p is the number of covariates}
}
\value{
a sequence of risk of rank correspond to each sample
}
\description{
list out all the possible variables, the order is according to P-value from small to large, and the cut-off-point has been optimal.
}
\details{
If the tree exists k terminal nodes, then the response 1 respresents the lowest risk and k represents the highest
}
\examples{
data(Lung,package="compound.Cox")
train_Lung=Lung[which(Lung[,"train"]==TRUE),] #select training data
t.vec=train_Lung[,1]
d.vec=train_Lung[,2]
x.mat=train_Lung[,-c(1,2,3)]
res=uni.tree(t.vec,d.vec,x.mat,P.value=0.01,d0=0.01,S.plot=FALSE,score=TRUE)
risk.classification(res,x.mat)
}
