% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni.logrank.R
\name{uni.logrank}
\alias{uni.logrank}
\title{Univariate selection under log-rank test}
\usage{
uni.logrank(t.vec, d.vec, X.mat)
}
\arguments{
\item{t.vec}{:Vector of survival times (time to either death or censoring)}

\item{d.vec}{:Vector of censoring indicators (1=death, 0=censoring)}

\item{X.mat}{:n by p matrix of covariates, where n is the sample size and p is the number of covariates}
}
\value{
A table contains an information of binary splitting, including P-value of the test, cut-off-point, sample sizes of two nodes, for each covariate

Pvalue: P-value of two sample logrank test

cut_off_point: to transfer x into binary case

left.sample.size: sample size of left child node

right.sample.size: sample size of right child node
}
\description{
List out all the possible variables. The cut-off-point has been optimal.
}
\details{
the output of the covariates has been ordered by the P-value from the smallest to the largest
}
\examples{
data(Lung,package="compound.Cox")
train_Lung=Lung[which(Lung[,"train"]==TRUE),] #select training data
t.vec=train_Lung[,1]
d.vec=train_Lung[,2]
x.mat=train_Lung[,-c(1,2,3)]
uni.logrank(t.vec,d.vec,x.mat)
}
