% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni.tree.R
\name{uni.tree}
\alias{uni.tree}
\title{A survival binary tree split by univariate selection and the statistical significant criterion.}
\usage{
uni.tree(
  t.vec,
  d.vec,
  X.mat,
  P.value = 0.01,
  d0 = 0.01,
  S.plot = FALSE,
  score = TRUE
)
}
\arguments{
\item{t.vec}{:Vector of survival times (time to either death or censoring)}

\item{d.vec}{:Vector of censoring indicators (1=death, 0=censoring)}

\item{X.mat}{:n by p matrix of covariates, where n is the sample size and p is the number of covariates}

\item{P.value}{:the threshold of P-value for stop splitting (stopping criterion)}

\item{d0}{:A positive constant to stabilize the variance of score statistics (Witten & Tibshirani 2010)}

\item{S.plot}{:call for plot the KM estimator for each split}

\item{score}{:TRUE = score test (Emura T et al. 2019); FALSE = log-rank test}
}
\value{
A nesting list of tree, containing inner node and terminal node information
}
\description{
Using a survival dataset to build a tree. The decision of splitting is using score test or log-rank test.
And the threshold of stop spltting can fixed through the argument.
This function also produce a relative risk rank between the terminal nodes.
}
\details{
the concept is generated by a function containing the function itself

For each split, we record three things, fist is the information of inner node which contains what to choose for splitting and what value the p-value the two sample test is. Second and third are the information of left and right children nodes.
If the children node is decided to become a inner node, then it genertae another two nodes;otherwise, it becomes a terminal node and record the information of sample size and the covariate space and the p-value it try to split but is failed
}
\examples{
data(Lung,package="compound.Cox")
train_Lung=Lung[which(Lung[,"train"]==TRUE),] #select training data
t.vec=train_Lung[,1]
d.vec=train_Lung[,2]
x.mat=train_Lung[,-c(1,2,3)]
uni.tree(t.vec,d.vec,x.mat,P.value=0.01,d0=0.01,S.plot=FALSE,score=TRUE)
}
\references{
Emura T, Matsui S, Chen HY (2019). compound.Cox: Univariate Feature Selection and Compound Covariate for Predicting Survival, Computer Methods and Programs in Biomedicine 168: 21-37.

Witten DM, Tibshirani R (2010) Survival analysis with high-dimensional covariates. Stat Method Med Res 19:29-51
}
